Remove aiodns requirement
aiohttp supports happyeyeballs so IPv6 workarounds should
no longer be needed.

Each poll was creating a new resolver, however aiodns
is designed to be created and reused over and over
since a single channel can support unlimited queries.

I considered only creating the resolver in an __init__,
but since the workaround was added in 2021 aiohttp gained
happyeyeballs support so it seemed that removing the workaround
was the best course.